prof can edit tranche

+++ app/Http/Controllers/EtudiantController.php
@@ -5,9 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\Filiere;
 use App\Models\Etudiant;
 use Illuminate\Http\Request;
-use App\Http\Requests\AddEtudiantRequest;
-use App\Http\Requests\UpdateEtudiantRequest;
 use Illuminate\Validation\Rule;
+use App\Http\Requests\AddEtudiantRequest;
+use Illuminate\Support\Facades\Validator;
+use App\Http\Requests\UpdateEtudiantRequest;
 
 class EtudiantController extends Controller
 {
@@ -24,13 +25,27 @@ class EtudiantController extends Controller
 
     public function destroy(Etudiant $etudiant)
     {
+        $validator = Validator::make([
+            'etudiant_cne' => $etudiant->id
+        ], [
+            'etudiant_cne' => 'required|unique:tranches',
+        ]);
+
+        if ($validator->fails()) {
+            return redirect()
+                        ->back()
+                        ->withErrors([
+                            'etudiant_id' => "le suppression de l'etudiant ($etudiant->prenom $etudiant->nom) ne peut pas etre effectue, car il a des tranches."
+                        ])
+                        ->withInput();
+        }
         $etudiant->delete();
 
         return redirect()->route('dashboard.etudiants');
     }
 
     public function update(Etudiant $etudiant, Request $request)
-    {
+    {   
         $validated = $request->validate(
             [
                 'prenom' => ['required', 'string', 'max:255'],
